refactor(server): extract inspection tools into src/server/handlers/inspection-tools.ts
get_variables, get_stack_trace, get_scopes, evaluate_expression,
get_source_context and get_local_variables move off the CallTool switch;
bodies verbatim with `this.` -> `ctx.` and `result = X; break;` ->
`return X;` (try-terminal and catch-branch-terminal assignments become
returns; the case's break followed the try/catch directly).

handleEvaluateExpression, handleGetSourceContext and handleGetLocalVariables
move too; the class keeps one-line private delegates as the test seams
(removed in PR 6).
